# specs/002-api-inference/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: AITRAF Demo Inference API
  version: 0.1.0
  description: FastAPI contract for demo video selection and inference.
servers:
  - url: http://localhost:8000
security:
  - bearerAuth: []
paths:
  /health:
    get:
      summary: Check API health
      security: []
      responses:
        "200":
          description: API is reachable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /demo-videos:
    get:
      summary: List current test-set demo videos
      parameters: []
      responses:
        "200":
          description: Flat demo video list for display and selection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DemoVideosResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /inference/trick-classification/{id}:
    post:
      summary: Run trick classification inference for a selected test-set row
      parameters:
        - $ref: "#/components/parameters/InferenceId"
      responses:
        "200":
          description: Trick classification prediction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InferenceResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/InvalidSelection"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /inference/trick-aqa/{id}:
    post:
      summary: Run trick AQA inference for a selected test-set row
      parameters:
        - $ref: "#/components/parameters/InferenceId"
      responses:
        "200":
          description: Trick AQA prediction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InferenceResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/InvalidSelection"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    InferenceId:
      name: id
      in: path
      required: true
      description: Stable row id returned by GET /demo-videos.
      schema:
        type: string
  responses:
    Unauthorized:
      description: Missing or invalid app token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Requested row id does not exist for this endpoint
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InvalidSelection:
      description: Request id is malformed or cannot be resolved by this endpoint
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServiceUnavailable:
      description: Required manifest, video, runtime config, processor, or model artifact is unavailable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    HealthResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]
    DemoVideosResponse:
      type: object
      required: [videos]
      properties:
        videos:
          type: array
          items:
            $ref: "#/components/schemas/DemoVideo"
    DemoVideo:
      type: object
      required: [id, video_id]
      properties:
        id:
          type: string
        video_id:
          type: string
        s3_path:
          type: string
        person:
          type: string
        trick:
          type: string
        execution_score:
          oneOf:
            - type: string
            - type: number
    InferenceResult:
      type: object
      required: [video_id, prediction, ground_truth, model]
      properties:
        video_id:
          type: string
        prediction:
          $ref: "#/components/schemas/PredictionResult"
        ground_truth:
          $ref: "#/components/schemas/DisplayResult"
        model:
          $ref: "#/components/schemas/ModelInfo"
    PredictionResult:
      type: object
      required: [label, confidence]
      properties:
        label:
          type: string
          description: User-facing prediction label or value, decoded from internal ids.
        confidence:
          type: number
          minimum: 0
          maximum: 1
          description: Model confidence normalized from 0 to 1.
    DisplayResult:
      type: object
      required: [label]
      properties:
        label:
          type: string
          description: User-facing label or value, decoded from internal ids.
    ModelInfo:
      type: object
      required: [kind]
      properties:
        kind:
          type: string
          description: Short demo-facing model kind, for example ordinal, video_mae, or fusion.
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string